# Parallelize phase linking for single-swath case in `single.py`