Improves performance for single burst (large swath, like NISAR) by adding block-level parallelism.

- Uses `ThreadPoolExecutor` since core JAX functions release GIL
- Refactored each block's processing to an inline `_process_block` function
- One tricky bit is the need for `read_lock` and `write_lock` to prevent race conditions on the background readers/writers

Closes #265 


Test case for a synthetic stack, shape `(15, 1997, 2138)`

Main branch: time for phase linking iteration: 51 seconds
```
dolphin run dolphin_config.yaml  114.60s user 8.00s system 203% cpu 1:00.17 total
```

PR Timings

- 1 worker: 55 seconds
```
dolphin run dolphin_config.yaml  112.15s user 7.90s system 192% cpu 1:02.33 total
```

- 3 workers: 24 seconds 
```
dolphin run dolphin_config.yaml  129.69s user 8.51s system 443% cpu 31.148 total
```

- 7 workers: 17 seconds
```
dolphin run dolphin_config.yaml  162.07s user 10.51s system 700% cpu 24.620 total
```

TODO:
- ~~Do we need to change the name of `n_parallel_bursts`? We could just use that for this argument (Answer: keeping the same for API compatibility. Documented this in the config string for `n_parallel_bursts`)~~
- ~~For a case where we pass 2 bursts, but set `n_parallel_bursts=7` (e.g.), should we then process each burst in parallel, and set the `max_workers` in single.py to `7 // 2 = 3` to run 3 blocks in parallel for both bursts?~~